rename setPostLanguage to getPostLanguage
